requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
Django==5.2
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
flake8==7.2.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
